--- TracksLyrics.Web/Requests/GetCurrentTrackLyricRequest.cs
@@ -1,6 +1,0 @@
-﻿namespace TracksLyrics.Web.Requests;
-
-public class GetCurrentTrackLyricRequest
-{
-    
-}
