+++ TracksLyrics.Web/Requests/GetTrackLyricRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace TracksLyrics.Web.Requests;
+
+public class GetTrackLyricRequest
+{
+    public string? Artist { get; set; }
+    public string? SongName { get; set; }
+}
